refactor: move pie chart contents to dedicated DataDisplayContents file.
