+++ src/Utils/ContentsConfig/DataDisplayContents.js
@@ -1,0 +1,12 @@
+export const pieChartContents = [
+  {
+    id: 1,
+    title: 'fill pie chart',
+    href: '#fill_pie_chart',
+  },
+  {
+    id: 2,
+    title: 'bordered pie chart',
+    href: '#bordered_pie_chart',
+  },
+];
