+++ src/Utils/ContentsConfig/InputContents.js
@@ -266,17 +266,3 @@ export const normalButtonContents = [
     href: '#buttons_shape',
   },
 ]
-
-// data display
-export const pieChartContents = [
-  {
-    id: 1,
-    title: 'fill pie chart',
-    href: '#fill_pie_chart',
-  },
-  {
-    id: 2,
-    title: 'bordered pie chart',
-    href: '#bordered_pie_chart',
-  },
-];
